refactor: RequestLogging 리팩터링
